feat: fixes routing rules tree view with better layout and node UI
