+++ ui/package.json
@@ -14,6 +14,7 @@
 	"dependencies": {
 		"@base-ui/react": "1.2.0",
 		"@bprogress/next": "3.2.12",
+		"@dagrejs/dagre": "3.0.0",
 		"@dnd-kit/helpers": "0.3.2",
 		"@dnd-kit/react": "0.3.2",
 		"@hookform/resolvers": "5.2.1",
